requirements.txt
passlib[argon2]
argon2-cffi
fastapi[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pydantic
alembic
uliweb-alembic
sqlmodel
pyjwt